move roi functions to a separate manager class
