+++ mmstudio/src/main/java/org/micromanager/internal/MMStudio.java
@@ -21,13 +21,9 @@ package org.micromanager.internal;
 import com.google.common.eventbus.Subscribe;
 import ij.IJ;
 import ij.ImageJ;
-import ij.ImagePlus;
 import ij.WindowManager;
-import ij.gui.Roi;
-import ij.gui.ShapeRoi;
 import ij.gui.Toolbar;
 import java.awt.Point;
-import java.awt.Rectangle;
 import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -37,7 +33,6 @@ import java.net.UnknownHostException;
 import java.util.*;
 import java.util.function.Function;
 import javax.swing.JOptionPane;
-import javax.swing.ToolTipManager;
 import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
 import mmcorej.CMMCore;
@@ -61,9 +56,7 @@ import org.micromanager.acquisition.internal.IAcquisitionEngine2010;
 import org.micromanager.alerts.AlertManager;
 import org.micromanager.alerts.internal.DefaultAlertManager;
 import org.micromanager.data.DataManager;
-import org.micromanager.data.Image;
 import org.micromanager.data.internal.DefaultDataManager;
-import org.micromanager.display.DataViewer;
 import org.micromanager.display.DisplayManager;
 import org.micromanager.display.internal.DefaultDisplayManager;
 import org.micromanager.events.AutofocusPluginShouldInitializeEvent;
@@ -142,6 +135,7 @@ public final class MMStudio implements Studio {
    private final MMSettings settings_ = new MMSettings();
    private MMCache cache_;
    private MMUIManager ui_;
+   private MMROIManager roi_;
    
    
    // MMcore
@@ -251,6 +245,7 @@ public final class MMStudio implements Studio {
       }
       
       // Start up multiple managers.  
+      roi_ = new MMROIManager(this);
       ui_ = new MMUIManager(this);
       userProfileManager_ = new UserProfileManager();       
       compatibility_ = new DefaultCompatibilityInterface(studio_);
@@ -492,12 +487,6 @@ public final class MMStudio implements Studio {
       UIMonitor.enable(OptionsDlg.isDebugLoggingEnabled(studio_));
    }
 
-   private void handleError(String message) {
-      live().setLiveModeOn(false);
-      JOptionPane.showMessageDialog(ui_.frame(), message);
-      core_.logMessage(message);
-   }
-
    /**
     * Spawn a new thread to load the acquisition engine jar, because this
     * takes significant time (TODO: Does it really, not that it is
@@ -522,138 +511,7 @@ public final class MMStudio implements Studio {
    public boolean getHideMDADisplayOption() {
       return AcqControlDlg.getShouldHideMDADisplay();
    }
-
-   public void setCenterQuad() {
-      ImagePlus curImage = WindowManager.getCurrentImage();
-      if (curImage == null) {
-         return;
-      }
-
-      Rectangle r = curImage.getProcessor().getRoi();
-      int width = r.width / 2;
-      int height = r.height / 2;
-      int xOffset = r.x + width / 2;
-      int yOffset = r.y + height / 2;
-
-      curImage.setRoi(xOffset, yOffset, width, height);
-      Roi roi = curImage.getRoi();
-      try {
-         app().setROI(updateROI(roi));
-      }
-      catch (Exception e) {
-         // Core failed to set new ROI.
-         logs().logError(e, "Unable to set new ROI");
-      }
-   }
-
-   public void setROI() {
-      ImagePlus curImage = WindowManager.getCurrentImage();
-      if (curImage == null) {
-         logs().showError("There is no open image window.");
-         return;
-      }
-
-      Roi roi = curImage.getRoi();
-      if (roi == null) {
-         // Nothing to be done.
-         logs().showError("There is no selection in the image window.\nUse the ImageJ rectangle tool to draw the ROI.");
-         return;
-      }
-      if (roi.getType() == Roi.RECTANGLE) {
-         try {
-            app().setROI(updateROI(roi));
-         }
-         catch (Exception e) {
-            // Core failed to set new ROI.
-            logs().logError(e, "Unable to set new ROI");
-         }
-         return;
-      }
-      // Dealing with multiple ROIs; this may not be supported.
-      try {
-         if (!(roi instanceof ShapeRoi && core_.isMultiROISupported())) {
-            handleError("ROI must be a rectangle.\nUse the ImageJ rectangle tool to draw the ROI.");
-            return;
-         }
-      }
-      catch (Exception e) {
-         handleError("Unable to determine if multiple ROIs is supported");
-         return;
-      }
-      // Generate list of rectangles for the ROIs.
-      ArrayList<Rectangle> rois = new ArrayList<>();
-      for (Roi subRoi : ((ShapeRoi) roi).getRois()) {
-         // HACK: just use the bounding box of each sub-ROI. Determining if
-         // sub-ROIs are rectangles is difficult (they "decompose" to Polygons
-         // once there's more than one at a time, so as far as I can tell we
-         // would have to test each angle of each polygon to see if it's
-         // 90 degrees and has the correct handedness), and this provides a
-         // good- enough solution for now.
-         rois.add(updateROI(subRoi));
-      }
-      try {
-         setMultiROI(rois);
-      }
-      catch (Exception e) {
-         // Core failed to set new ROI.
-         logs().logError(e, "Unable to set new ROI");
-      }
-   }
-
-   /**
-    * Adjust the provided rectangular ROI based on any current ROI that may be
-    * in use.
-    */
-   private Rectangle updateROI(Roi roi) {
-      Rectangle r = roi.getBounds();
-
-      // If the image has ROI info attached to it, correct for the offsets.
-      // Otherwise, assume the image was taken with the current camera ROI
-      // (which is a horrendously buggy way to do things, but that was the
-      // old behavior and I'm leaving it in case there are cases where it is
-      // necessary).
-      Rectangle originalROI = null;
-
-      DataViewer viewer = displays().getActiveDataViewer();
-      if (viewer != null) {
-         try {
-            List<Image> images = viewer.getDisplayedImages();
-            // Just take the first one.
-            originalROI = images.get(0).getMetadata().getROI();
-         }
-         catch (IOException e) {
-            ReportingUtils.showError(e, "There was an error determining the selected ROI");
-         }
-      }
-
-      if (originalROI == null) {
-         try {
-            originalROI = core().getROI();
-         }
-         catch (Exception e) {
-            // Core failed to provide an ROI.
-            logs().logError(e, "Unable to get core ROI");
-            return null;
-         }
-      }
-
-      r.x += originalROI.x;
-      r.y += originalROI.y;
-      return r;
-   }
-
-   public void clearROI() {
-      live().setSuspended(true);
-      try {
-         core_.clearROI();
-         cache().refreshValues();
-
-      } catch (Exception e) {
-         ReportingUtils.showError(e);
-      }
-      live().setSuspended(false);
-   }
-
+   
    @Override
    public CMMCore core() {
       return core_;
@@ -882,7 +740,6 @@ public final class MMStudio implements Studio {
       if (ui_.frame() != null) {
          ui_.frame().savePrefs();
       }
-
       // NOTE: do not save auto shutter state
       if (afMgr_ != null && afMgr_.getAutofocusMethod() != null) {
          profile().getSettings(MMStudio.class).putString(AUTOFOCUS_DEVICE, afMgr_.getAutofocusMethod().getName());
@@ -1099,13 +956,6 @@ public final class MMStudio implements Studio {
       }
    }
 
-   public void setMultiROI(List<Rectangle> rois) throws Exception {
-      live().setSuspended(true);
-      core_.setMultiROI(rois);
-      cache().refreshValues();
-      live().setSuspended(false);
-   }
-
    public void setAcquisitionEngine(AcquisitionWrapperEngine eng) {
       acqEngine_ = eng;
    }
@@ -1291,6 +1141,10 @@ public final class MMStudio implements Studio {
 
    public MMUIManager uiManager() {
       return ui_;
+   }
+   
+   public MMROIManager roiManager() {
+      return roi_;
    }
       
    public class MMSettings {
